+++ api-js/src/enums/index.js
@@ -1,6 +1,6 @@
 export * from './dkim-order-field'
 export * from './dkim-result-order-field'
-export * from './dmarc-order'
+export * from './dmarc-order-field'
 export * from './dmarc-summary-order-field'
 export * from './domain-order-field'
 export * from './https-order-field'
